binupdate improvements
- new safeDownload() in helperFunctions
- introduced showProgress argument for emulator downloads
- modified binupdate.sh to use showProgress
- modified emulator/tools install functions to use safeDownload()/showProgress and return proper exit code (CemuNative, Cemu, mGBA, PCSX2-QT, Ryujinx, Vita3K, Xenia, Yuzu, ESDE, SRM)
- installEmuAI() and installEmuBI() are now using new safeDownload() and return proper exit code
- installEmuAI() and installEmuBI() can now write last version info (new arguments)
- yuzu launcher: Added ~/emudeck/yuzu.noupdate to skip yuzu update, zenity progress bar and zenity error dialog

+++ functions/EmuScripts/emuDeckVita3K.sh
@@ -14,9 +14,13 @@ Vita3K_cleanup(){
 #Install
 Vita3K_install(){
     echo "Begin Vita3K Install"
-    installEmuBI "Vita3K"  "$(getReleaseURLGH "Vita3K/Vita3K" "ubuntu-latest.zip")" "Vita3K" "zip"
-    unzip -o "$HOME/Applications/Vita3K.zip" -d "$Vita3K_emuPath" && rm -rf "$HOME/Applications/Vita3K.zip"
-    chmod +x "$Vita3K_emuPath/Vita3K"
+    local showProgress="$1"
+    if installEmuBI "Vita3K" "$(getReleaseURLGH "Vita3K/Vita3K" "ubuntu-latest.zip")" "Vita3K" "zip" "$showProgress"; then
+        unzip -o "$HOME/Applications/Vita3K.zip" -d "$Vita3K_emuPath" && rm -rf "$HOME/Applications/Vita3K.zip"
+        chmod +x "$Vita3K_emuPath/Vita3K"
+    else
+        return 1
+    fi
 }
 
 #ApplyInitialSettings
