+++ functions/EmuScripts/emuDeckXenia.sh
@@ -16,23 +16,28 @@ Xenia_cleanup(){
 Xenia_install(){
 	local version
 	version=$1
+	local showProgress="$2"
 
 	if [[ "$version" == "master" ]]; then
 		Xenia_releaseURL="$Xenia_releaseURL_master"
 	else
 		Xenia_releaseURL="$Xenia_releaseURL_canary"
 	fi
-
+    local name="$Xenia_emuName-$version"
 	
 	setMSG "Installing Xenia $version"		
 
 	#need to look at standardizing exe name; or download both?  let the user choose at runtime?
-	curl -L "$Xenia_releaseURL" --output "$romsPath"/xbox360/xenia.zip 
-	mkdir -p "$romsPath"/xbox360/tmp
-	unzip -o "$romsPath"/xbox360/xenia.zip -d "$romsPath"/xbox360/tmp
-	rsync -avzh "$romsPath"/xbox360/tmp/ "$romsPath"/xbox360/
-	rm -rf "$romsPath"/xbox360/tmp
-	rm -f "$romsPath"/xbox360/xenia.zip 	
+	#curl -L "$Xenia_releaseURL" --output "$romsPath"/xbox360/xenia.zip 
+    if safeDownload "$name" "$Xenia_releaseURL" "$romsPath"/xbox360/xenia.zip" "$showProgress"; then
+		mkdir -p "$romsPath"/xbox360/tmp
+		unzip -o "$romsPath"/xbox360/xenia.zip -d "$romsPath"/xbox360/tmp
+		rsync -avzh "$romsPath"/xbox360/tmp/ "$romsPath"/xbox360/
+		rm -rf "$romsPath"/xbox360/tmp
+		rm -f "$romsPath"/xbox360/xenia.zip
+	else
+		return 1
+	fi
 
 	cp "$EMUDECKGIT/tools/launchers/xenia.sh" "${toolsPath}/launchers/xenia.sh"
 	sed -i "s|/run/media/mmcblk0p1/Emulation/tools|${toolsPath}|g" "${toolsPath}/launchers/xenia.sh"
@@ -41,8 +46,7 @@ Xenia_install(){
 #	if [[ "$launchLine"  == *"PROTONLAUNCH"* ]]; then
 #		changeLine '"${PROTONLAUNCH}"' "$launchLine" "${toolsPath}/launchers/xenia.sh"
 #	fi
-	chmod +x "${toolsPath}/launchers/xenia.sh"
-	
+	chmod +x "${toolsPath}/launchers/xenia.sh"	
 
 	createDesktopShortcut   "$HOME/.local/share/applications/xenia.desktop" \
 							"Xenia (Proton)" \
